requirements.txt
# API Clients
spotipy>=2.19.0
pylast>=5.0.0

# Database
sqlalchemy>=1.4.0

# CLI
click>=8.0.0

# Configuration
python-dotenv>=0.19.0

# Development dependencies (uncomment as needed)
# pytest>=6.0.0
# pytest-cov>=2.12.0
# black>=21.5b2
# isort>=5.9.0
# flake8>=3.9.0
# pre-commit>=2.13.0
